+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Pass infrastructure machinepool object labels to IAM role creation.
+
 ## [1.0.0] - 2025-02-19
 
 ### Changed